+++ passkit_ui/lib/src/coupon.dart
@@ -3,6 +3,7 @@ import 'package:passkit/passkit.dart';
 import 'package:passkit_ui/passkit_ui.dart';
 import 'package:passkit_ui/src/theme/coupon_theme.dart';
 import 'package:passkit_ui/src/widgets/header_row.dart';
+import 'package:passkit_ui/src/widgets/strip_image.dart';
 
 /// A coupon looks like the following:
 ///
@@ -20,8 +21,6 @@ class Coupon extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final devicePixelRatio = MediaQuery.devicePixelRatioOf(context);
-
     final theme = Theme.of(context).extension<CouponTheme>()!;
     final coupon = pass.pass.coupon!;
 
@@ -48,9 +47,7 @@ class Coupon extends StatelessWidget {
               Stack(
                 children: [
                   if (pass.strip != null)
-                    Image.memory(
-                      pass.strip!.forCorrectPixelRatio(devicePixelRatio),
-                    ),
+                    StripImage(image: pass.strip, type: PassType.coupon),
                   Column(
                     crossAxisAlignment: CrossAxisAlignment.stretch,
                     children: [
@@ -74,21 +71,15 @@ class Coupon extends StatelessWidget {
               ),
               const SizedBox(height: 16),
               _AuxiliaryRow(
+                auxiliaryRow: coupon.secondaryFields ?? [],
                 passTheme: theme,
-                auxiliaryRow: [
-                  ...?coupon.secondaryFields,
-                  ...?coupon.auxiliaryFields,
-                ],
+              ),
+              _AuxiliaryRow(
+                auxiliaryRow: coupon.auxiliaryFields ?? [],
+                passTheme: theme,
               ),
               const SizedBox(height: 16),
               const Spacer(),
-              if (pass.footer != null)
-                Image.memory(
-                  pass.footer!.forCorrectPixelRatio(devicePixelRatio),
-                  fit: BoxFit.contain,
-                  width: 286,
-                  height: 15,
-                ),
               if ((pass.pass.barcodes?.firstOrNull ?? pass.pass.barcode) !=
                   null)
                 PasskitBarcode(